chore: refactor things into separate components
